Move search index generator command logic to task
